\name{ProDenICA-internal}
\title{Internal ProDenICA functions}
\description{Internal ProDenICA functions}
\alias{d.gaussmix}
\alias{dmix.dexp}
\alias{modulus}
\alias{r.gaussmix}
\alias{rmix.dexp}
\author{Trevor Hastie}
\details{These are not intended for use by users. Thes functions are
  used by \code{djordan} and \code{rjordan}
}
\keyword{internal}
